python-backend/requirements.txt
# Chimaridata Python Backend Requirements
# These are the CORE dependencies required for the backend to run.
# Install with: pip install -r requirements.txt
#
# Optional features:
#   PII detection:  pip install -e ".[pii]"
#   Dev tools:     pip install -e ".[dev]"
#   Everything:    pip install -e ".[all]"

# ============================================================================
# Core Framework
# ============================================================================
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# ============================================================================
# LangChain & LangGraph
# ============================================================================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0

# ============================================================================
# AI/LLM Providers
# ============================================================================
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.2
langchain-google-genai>=1.0.1

# ============================================================================
# Vector Stores
# ============================================================================
faiss-cpu>=1.13.0
pgvector>=0.2.4

# ============================================================================
# Data Processing
# ============================================================================
pandas>=2.1.4
numpy>=1.26.0

# ============================================================================
# Database
# ============================================================================
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.7
alembic>=1.12.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.1
pytz>=2023.3
python-dateutil>=2.8.2
email-validator>=2.1.0

# ============================================================================
# Real-time
# ============================================================================
websockets>=12.0

# ============================================================================
# Payment Processing
# ============================================================================
stripe>=7.0.0

# ============================================================================
# Testing (basic)
# ============================================================================
pytest>=7.4.3
pytest-asyncio>=0.21.0
